pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.10.0",
    "attrs>=25.4.0",
    "django>=5.2.7",
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "django-unfold>=0.67.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "jsonschema>=4.25.1",
    "jsonschema-specifications>=2025.9.1",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
]
